\name{dreadds}
\alias{dreadds}
\docType{package}
\title{
Dynamic Range Evolution and Diversification simulation with dynamic speciation model
}
\description{
FIXME
Implements the Dynamic Range Evolution and Diversification simulation with dynamic speciation model

}
\usage{
dreadds()
}
\examples{
\dontrun{
dreadds()
}
}
