\name{dreadds}
\alias{dreadds}
\docType{package}
\title{
Dynamic Range Evolution and Diversification simulation with dynamic speciation model
}
\description{
Implements the Dynamic Range Evolution and Diversification simulation with dynamic speciation model

}
\usage{
dreadds(path_to_config_file)
}
\examples{
\dontrun{
dreadds("inputs/example.conf")
}
}
